path: guard all entry points with this.path and fs dependency check
this.path and fs are suggested packages.  With recent additions, fs
and this.path are used in more spots, and some of them aren't guarded.

Go with the simpler approach of guarding each exported function with a
check for fs and this.path, which makes it easier to see that
everything is covered.  It has the downside that some functionality
doesn't technically need fs, but fs is very likely brought into
projects anyway.

(It's probably worth considering just adding fs and this.path as hard
dependencies.)