path: pass empty environment to this.path functions
Several spots pass the result of caller_env to this.path functions.
The intention is to pass the caller environment of the entry point
invoked by the _user_.  However, that's not the case in several spots
because caller_env is called within an _internal_ function.

We could plumb the caller_env result from the entry point.  However,
it's not clear that's worth the complication.  I've been unable to
find any use cases in the project context that would depend on passing
an environment to this.path's envir argument.

For now, pass an empty environment to make it clear we're not relying
on the envir lookup.  If a use case comes up, we can adjust for it
later.