tf_options: label non-existing paths
mrg_script, tables_to, and figures_to warn but still set the option if
a path does not exist.  Teach tf_options to flag these cases.

Test tf_options by including it in the sourced scripts.  (As with the
other path functions, it needs to be put in a file that issourced so
that this.path returns the temporary directory as the project rather
than the source repository.)