path: use local_dir to set directory for tables_to/figures_to tests
The tables_to/figures_to tests wrap the source call within a with_dir
call that changes to the script directory so that this.path returns
the test project (set up in a temporary directory) rather than
top-level of the source code repository.

That directory is the relevant one for all the checks in this block,
so set that directory for the entire block with local_dir.