path: fix 'R CMD check' test failures
When running under the temporary directory for R CMD check, two tests
that trigger an mrg_script call fail:

 * test-path.R:57:3 (set script name option)
 * test-path.R:155:3 (warn when path doesn't exist)

In both cases, this.path::this.proj aborts with

  no root directory found in "{tmppdir}/mrgmisc.Rcheck/tests/testthat"
  or its parent directories

devtools::test passes because this.path::this.proj returns the
top-level of the source repository.  (I'm not sure why the GitHub
Actions R CMD check passes.)

Make this.path::this.proj work under R CMD check by moving these calls
to a sourced file, as is already done for other cases.