tables_to, figures_to: expand created path with path_real too
When the passed path is an existing directory, tables_to and
figures_to expand it with fs::path_real, which requires the path to
exist.  bad_path used to warn on a non-existing path, so _either_ the
bad_path or path_real were executed.

As of 7e7a17f, bad_path either aborts or creates the directory,
depending on whether create=TRUE is passed.  So, if bad_path doesn't
abort, path_real should be called.